Nothing relevant. Only a set of castings to have a clean compile on Windows. The cl.exe compiler is pretty good at complaining about any kind of non explicit cast.
This commit was SVN r12207.

+++ orte/mca/pls/base/pls_base_dmn_registry_fns.c
@@ -62,7 +62,7 @@ int orte_pls_base_store_active_daemons(opal_list_t *daemons)
     int rc, i, num_daemons;
     
     /* determine the number of daemons */
-    num_daemons = opal_list_get_size(daemons);
+    num_daemons = (int)opal_list_get_size(daemons);
 
     if (0 == num_daemons) {
         return ORTE_SUCCESS;
